requirements.txt
customtkinter
openpyxl
matplotlib
cryptography
pymysql
python-dotenv